front_end/requirements.txt
reflex>=0.5.0
pandas>=2.0.3
google-cloud-storage>=2.10.0
google-auth>=2.23.0
plotly>=5.17.0
pytz>=2023.3
python-dateutil>=2.8.2
